Upload datasets documents (#5)
* interface (#1)

upload the current app of the final project

* obsolete code (#3)

datasets treatment code, clusterization and visualization graphs

* upload datasets documents

Uploaded all the datasets that I retrieve from the web and that I transform with the code that I just uploaded too. All the documents are updated since now, and are also uploaded to the repository (highlight the _Memory document that contains the description of all the work done all the previous days --> its important to keep updating it).