Weighted clustering model (#10)
* keywords without threshold

* name and new dow

changed the name of the obsolete matrix of researchers and keywords (the one with threshold = 10 due to that now I am going to try the results with weights, and I need all the words)

* weighted script creation

for kmeans and to try the symmetric penalization

* inconclusive weighted clustering tests

after testing various weighted clustering methods, the results improved in some situations, but generally remained the same or worse. this approach is also not suitable for the data.