clustering word embedding finished (#13)
* interface (#1)

upload the current app of the final project

* obsolete code (#3)

datasets treatment code, clusterization and visualization graphs

* upload datasets documents (#6)

Uploaded all the datasets that I retrieve from the web and that I transform with the code that I just uploaded too. All the documents are updated since now, and are also uploaded to the repository (highlight the _Memory document that contains the description of all the work done all the previous days --> its important to keep updating it).

* Clusterization model (#8)

* clustering

Upgrade of the clustering model and the creation of its result

* word of parameters

added a word to takes notes about the tries done in the clustering model

* new test of the model

* model upgrade

documentation upgraded and code developed in a proper way

* Delete _Clusterization/Investigadores y keywords/20250924 Modelo de clusterización-Kmeans.R

* Delete _Clusterization/Investigadores y keywords/Pruebas clustering.docx

* Clusterization model (#9)

* clustering

Upgrade of the clustering model and the creation of its result

* word of parameters

added a word to takes notes about the tries done in the clustering model

* new test of the model

* model upgrade

documentation upgraded and code developed in a proper way

* new models

and update of doc

* Weighted clustering model (#10)

* keywords without threshold

* name and new dow

changed the name of the obsolete matrix of researchers and keywords (the one with threshold = 10 due to that now I am going to try the results with weights, and I need all the words)

* weighted script creation

for kmeans and to try the symmetric penalization

* inconclusive weighted clustering tests

after testing various weighted clustering methods, the results improved in some situations, but generally remained the same or worse. this approach is also not suitable for the data.

* testing of word embeddings clustering (#12)

creation and development of the use of word embeddings for describe the structure of the data (by clusterization, as before)

* clustering WE finished

i obtained proper results (S=0.6), so i consider as finished this section

* delete unnecesary csv

about the results of the clusters that i am not going to use finally