update _memory (#25)
* corrections from the meeting (#15)

* update of the _Memory (#17)

* deep investigation of the clusters (#19)

obtaning the researchers and keywords from each cluster of UMAP clustering and non-UMAP clustering. there's also added a draft of the conclusions obtained by the comparisons of both clustering methods results. there's remaining the visualization of the results to achieve proper comparisons

* fixing branch problems

* Comparation definitive clusterings (#21)

* deep investigation of the clusters

obtaning the researchers and keywords from each cluster of UMAP clustering and non-UMAP clustering. there's also added a draft of the conclusions obtained by the comparisons of both clustering methods results. there's remaining the visualization of the results to achieve proper comparisons

* finishing comparion clustering models

* title and description

first ideas for the title and description of the TFG

* update _memory (#24)