Clust without locl (#28)
* embeddings wo locl (1)

* preparation new clustering

preparing clustering without localization terms

* datasets to clusterize

creation and concatenation of the datasets which are going to be used to create the new clustering models (the ones without the words that refer to location). In that sense, the new dataset of embeddings for searchers are made by the mean of all the not-location words.